feat: Pipe CSV contents to docker
replace volume with piped stdin

Previously, only a filename was being passed and so the file needed to
actually be accessible to the container via mounted volume. Now instead,
since the CSVs are expected to be tiny, the raw contents of the file can
be piped in to the container thus eliminating the need for any volumes
or for moving files around on the host computer.

fix: actually need a volume for outputs

include parsed totals in output

display data

sort values by time

add runner script

output spent amount